a2xx: add initial register definitions
Not as complete as freedreno_a2xx_reg.h, but as complete as the previous
hand-written parsing in cffdump.  So we can at least merge the envytools
changes to master and dump the hand-written parsing.  A bit more is
needed before we can attempt to replace the hand written header file.