rnn: add headergen2
Modified version of headergen which generates inline fxns for setting
register/bitfield values and 'C' enums for enum values, in order to have
some sort of type safety.  The inline fxns also have asserts for
min/max/align to preserve the error checking.

Possibly the shift handling is wrong for nested bitsets.. I'm not really
using that feature for adreno.