squash! WIP: add rnndb for MDP40
Remove variants for now.. possibly we want variants for revisions of
mdp4, but different db files for mdp3, mdss, etc.  Not sure how similar
the different display controller blocks are, but I think it makes more
sense to try and deal with variants once the differences are better
understood.

And rename mdp->mdp4