rnn: add support for irregular arrays
Add support for "arrays" with non-regular base addresses.  Sometimes hw
designers put identical blocks of registers in goofy places.  But for
writing a driver or looking at decoded traces, it is easier to think of
it as conceptually an array.

(TODO: headergen is not updated for this yet, since I don't use it..)