a4xx: start adding a4xx registers
A handful we can figure out from downstream kgsl kernel driver, which is
starting to receive A4XX support.  Not as many useful hints as previous
generations.  But it's a starting point.  Probably some of the registers
are basically same as A3XX, but just different offset.  But there look
to be a lot new (seems like we get geometry shaders, hull shaders, etc
\o/).  Currently there are some easy-to-fill-in-blanks in the register
names.