rnndb: Update registers / add additional enums for devcoredump
We are moving the PDC registers to their own region in the DT and
the offsets need to be changed accordingly. Similarly move GX_DBGC
to its own domain.

Also, add the enums and other bits needed to support the expanded
devcoredump for a6xx and fix a transposed bit value for GMU.