afuc: print instructions as literal if pad bits
If we see unknown pad bits, the disassembler should print instructions
as literal (with best attempt decoding as a comment), so that we can
round-trip disasm -> asm.  Eventually when we figure out what the extra
pad bits mean, we wouldn't need this.

Note this means moving around where flush (f) flag is printed, so it
doesn't appear before the literal.